+++ packages/client/CHANGELOG.md
@@ -20,6 +20,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Remove deprecated:
+  - `gasPriceStrategy` config option from `contracts.ethereumNetworks`
+  - method parameter of `.waitForStorage()`
+  - `STREAM_CLIENT_DEFAULTS` constant
+  - `ConfigTest` constant
+  - `StrictStreamrClientConfig` TypeScript interface
+  - `NetworkNodeConfig` TypeScript interface
+- Remove `BigNumber` class export (was only used in `gasPriceStrategy`)
+
 ### Fixed
 
 ### Security